Start creation of a checktag command.
This command creates tag provenance (new!) and VSAs for those tags.

There's more work to do to fully iron this out but it's ready for
some testing.

This was a fairly large refactor that involved generalizing the
GitHubConnection class so that it doesn't assume there are branches and
being able to mock the bnd verifier to make unit tests possible/easier.

Signed-off-by: Tom Hennen <tomhennen@google.com>