Push the policy time forward again.
There's an issue where we don't look at all the available
provenance just the first one we find.  So even though
prior commits have 'good' provenance that would meet this
policy they also have 'bad' provenance that doesn't. That
bad provenance is found 'first' which means we fail the
policy check.  We could fix that, but it's tricky. Better
to get things working again and we can resolve later.

Signed-off-by: Tom Hennen <tomhennen@google.com>