Generalize policy package (#233)
* Split prov predicate to own pkg

This commit splits out the provenance predicate to its own
package in preparation to generate it from protobuf definitions.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* PolicyEval: Use generic models

This commit modifies the policy package to take the generic
models instead of the github connection for the policy operations.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* sourcetool: Drop ghcontrol dep

Witht this commit, the sourcetool package no longer requires the
ghcontrll package as it is now used only through the github backend.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* CLI: Use new policy interfaces

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* Regenerate fakes

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* Go mod tidy

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

---------

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>