Update verifier paths, bridge ID migration (#256)
* Fix tag check updating URIs

This commit fixes the broken tag verification by updating the paths
and URIs for the verification IDs to those of the new repos.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* Use constants in default signer IDs

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

* COMPAT: Support old+new signers

This commits adds a compatibility hack to support both the old and new
actions repository signer identities while we migrate to the new
source-actions repos.

This commit is inteded to be reverted once all repos have signed their
VSAs using the new identity. For more see:

https://github.com/slsa-framework/slsa-source-poc/issues/255

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>

---------

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@carabiner.dev>